Consolidate http headers (#113)
* Consolidate http headers

* tidy up comments
