+++ cmd/flags/http.go
@@ -34,7 +34,6 @@ var allowedHTTPMethods = map[string]interface{}{
 
 // HTTP stores flags related to HTTP requests.
 type HTTP struct {
-	Headers  stringArray
 	Requests stringArray
 }
 
@@ -43,12 +42,7 @@ func (h *HTTP) String() string {
 }
 
 func (h *HTTP) initFlags() {
-	flag.Var(&h.Headers, "http-headers", "HTTP header to be sent with warm up requests.")
 	flag.Var(&h.Requests, "http-requests", `HTTP request to be sent. Request is in '<http-method>:<path>[:body]' format. E.g. post:/ping:{"key":"value"}`)
-}
-
-func (h *HTTP) getWarmupHTTPHeaders() map[string]string {
-	return toHeaders(h.Headers)
 }
 
 func (h *HTTP) getWarmupHTTPRequests() ([]http.Request, error) {
