+++ pkg/util/util_test.go
@@ -12,11 +12,12 @@
 //See the License for the specific language governing permissions and
 //limitations under the License.
 
-package flags
+package util
 
 import (
-	"github.com/stretchr/testify/assert"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func Test_ToHeaders(t *testing.T) {
@@ -26,7 +27,7 @@ func Test_ToHeaders(t *testing.T) {
 		"Host: localhost",
 	}
 
-	headers := toHeaders(headersFlag)
+	headers := ToHeaders(headersFlag)
 
 	assert.Equal(t, 2, len(headers))
 	assert.Equal(t, "application/json", headers["Content-Type"])
@@ -39,7 +40,7 @@ func Test_HeadersWithoutSeparatorToHeaders(t *testing.T) {
 		"No separator",
 	}
 
-	headers := toHeaders(headersFlag)
+	headers := ToHeaders(headersFlag)
 
 	assert.Equal(t, 1, len(headers))
 	value, ok := headers["No separator"]
@@ -53,7 +54,7 @@ func Test_HeadersWithMultipleSeparatorsToHeaders(t *testing.T) {
 		"Cookie: some:strange:cookie",
 	}
 
-	headers := toHeaders(headersFlag)
+	headers := ToHeaders(headersFlag)
 
 	assert.Equal(t, 1, len(headers))
 	assert.Equal(t, "some:strange:cookie", headers["Cookie"])
