+++ pkg/warmup/warmup.go
@@ -20,6 +20,7 @@ import (
 	"mittens/pkg/grpc"
 	"mittens/pkg/http"
 	"mittens/pkg/safe"
+	"mittens/pkg/util"
 	"sync"
 	"time"
 )
@@ -30,9 +31,8 @@ type Warmup struct {
 	MaxDurationSeconds       int
 	Concurrency              int
 	HttpRequests             chan http.Request
-	HttpHeaders              map[string]string
+	HttpHeaders              []string
 	GrpcRequests             chan grpc.Request
-	GrpcHeaders              []string
 	RequestDelayMilliseconds int
 }
 
@@ -45,7 +45,7 @@ func (w Warmup) Run(hasHttpRequests bool, hasGrpcRequests bool, requestsSentCoun
 	if hasGrpcRequests {
 		// connect to gRPC server once and only if there are gRPC requests
 		log.Print("gRPC client connecting...")
-		connErr := w.Target.grpcClient.Connect(w.GrpcHeaders)
+		connErr := w.Target.grpcClient.Connect(w.HttpHeaders)
 
 		if connErr != nil {
 			log.Printf("gRPC client connect error: %v", connErr)
@@ -54,7 +54,7 @@ func (w Warmup) Run(hasHttpRequests bool, hasGrpcRequests bool, requestsSentCoun
 				log.Printf("Spawning new go routine for gRPC requests")
 				wg.Add(1)
 				go safe.Do(func() {
-					w.GrpcWarmupWorker(&wg, w.GrpcRequests, w.GrpcHeaders, w.RequestDelayMilliseconds, requestsSentCounter)
+					w.GrpcWarmupWorker(&wg, w.GrpcRequests, w.HttpHeaders, w.RequestDelayMilliseconds, requestsSentCounter)
 				})
 			}
 		}
@@ -65,7 +65,7 @@ func (w Warmup) Run(hasHttpRequests bool, hasGrpcRequests bool, requestsSentCoun
 			log.Printf("Spawning new go routine for HTTP requests")
 			wg.Add(1)
 			go safe.Do(func() {
-				w.HTTPWarmupWorker(&wg, w.HttpRequests, w.HttpHeaders, w.RequestDelayMilliseconds, requestsSentCounter)
+				w.HTTPWarmupWorker(&wg, w.HttpRequests, util.ToHeaders(w.HttpHeaders), w.RequestDelayMilliseconds, requestsSentCounter)
 			})
 		}
 	}
